// src/app/models/response/loginResponse.model.ts
export class LoginResponseModel{
    empId: string;
    empName: string;
    mobileNo: string;
    emailId: string;
    role: string;
    empRole: string;
    rmId: string;
    terName: string;
    terCode: string;
    orgName: string;
    locName: string;
    locCode: string;
    gradeTitle: string;
    departName: string;
    busUnit: string;
    designation: string;
}